requirements.txt
djangorestframework-3.13.1
Django==4.1
Pillow==9.2.0
python-dotenv==0.20.0
